+++ NEWS.md
@@ -1,6 +1,7 @@
 # BGData 2.2.0.9000
 
 - `preprocess()`: Add `nCores` parameter.
+- `findRelated()`: Map indices to sample names for matrix inputs.